feat: Added Breadcrumbs component and content to "Hot" page
Added a new Breadcrumbs component to improve navigation within the application.
This component displays the user's current location and provides the ability to navigate back through the navigation history.
Added MenuBlock component to manage FoodCard components display.
Additionally, content has been added to the "Hot" page, including a list of dishes from the "Hot Dishes" category.
Each dish card is now interactive, and clicking on it will redirect the user to the FoodPage component which is not ready yet.